Added region information statistics and commandline options.
Added the option --region-information that will output the defintion
of the region used for imagaging, and basic statistics (Number of particles
and number of imaged particles) per region.  THere is --region-definition
option to override the region definition and provide one yourself.

The next step is adding region output to each particle, and some more
testing.

This information combined with the analysed volume that is already reported
can be used to calculate an imaged volume. NOTE: If this imaged volume is
valid depends on several other settings in CytoUsb, including the actual
imaging mode.