Add support for "Target All" imaging.
When using target all, there is no regioninformation stored in the
file.  To handle this we create a default setlist, with only a single
set, which is all particles.  This way it can be treated the same as
any other target range.

The only difference is that we do not explicitely list the name of
the default set for each particle, because a particle is allways in the
default set.